Code consistency: Include file in repository, even though used only by Pro
This file was previously excluded from the Permissions Free release to avoid misunderstanding.  It is only used by Pro, and applies only to pre-2012 installations of the original pre-release PressPermit beta (which was never a free release on wordpress.org).  It is being added to the repository for vendor library usage by Pro, and release process consistency.