Authors integration: Require Authors version 3.8.0 or later
Don't apply JOIN clause to the posts query (to credit multiple authors the same as post_author) unless Authors version is 3.8.0 or higher.

This ensures mapped Authors can not be created or edited to have a slug not equal to user_nicename.  It also eliminates several bugs and quirks which would cause unexpected Permissions filtering results.