Move vendor autoload call to out of the plugins_loaded action callback
This is required because our shared libraries use that same hook to register and initialize their instances. Those libraries use Composer's file autoload feature to trigger the scripts that register the hooks, and those hooks should be already registered when the plugins_loaded action runs, honoring the correct priority order to be sure the libs are available when the main callback that loads the plugin is finally triggered.

Update libraries to replace pimple and psr/container with prefixed libs.