Fix inconsistencies in update script triggering; introduce installation log
Prevent Core update scripts from being skipped when a Pro package is installed for the first time.

Also prevent hypothetical installation scripts in Pro modules from being skipped if a Pro package is installed for the first time but contains a Core library with the same version that was previously installed directly.

Introduce an installation log to display a version history (with install date) in Permissions > Settings > Install.