Edit Permissions: Resolve previously noted issues with new current_exceptions table
https://github.com/publishpress/PublishPress-Permissions/pull/1555#issuecomment-2828523024

"I think using icons to mark Enabled / Blocked / Limit is OK. However, can we display a guide under each collapsible section?"

    Checkmark / X / !,
    ... = assigned for categories and sub-categories
    any other captions.

"Can you structure this with invisible columns?"

"The key caption "* = Permissions redundant" should not be displayed unless a listed item is so marked."

"In my test case, the Permissions count for Read-Posts does not match the number of items listed."

I've also clarified the captions of the main sections to indicate the post type or taxonomy that Permissions are assigned to. Within that are subsections grouped by what access the Permissions alter:

* Category Permissions
-- Read Posts
-- Edit Posts
-- Read Pages

* Post Tag Permissions
-- Read Posts
-- Manage Post Tags

* Page Permissions
-- Read Pages
-- Edit Pages
-- Create Revision Pages

Both the main sections and subsections are collapsible. Styling is updated for better color contrast and spacing.

To eliminate the possibility of a mismatch between calculated Permissions item count and actual display, those counts are now updated by javascript based on the logged output.

The status caption of the current_roles table is fixed for Statuses plugin compat.

The Modification Type and Status icon is moved to the left of the item name for visual alignment.

The width of each column is optimized for visual alignment.

Tool tip text is revised for brevity and context relevance:
- Open tool tip with "Enable" / "Blockage" / "Limitation"
- Name actual access type modified
- Name post type or taxonomy the Permission is assigned to (instead of "item")
- If the Permission is assigned for sub-posts, say so
- Use a dashicon to clarify assignment for sub-posts / sub-terms (with color matching previous captioning)
- Show a key for Modification Type and Assignment type in the bottom of each main section
- Use a dashicon and tooltip to flag and explain redundant assignments

Fixes #1319